+++ .env_example.json
@@ -1,4 +1,7 @@
 {
-    "WALLET_PRIVATE_KEY": "",
-    "INFURA_NODE_ENDPOINT": ""
+    "REMOTE_TEST_PRIVATE_KEY1": "",
+    "REMOTE_TEST_PRIVATE_KEY2": "",
+    "INFURA_NODE_ENDPOINT": "",
+    "ADDRESS1": "",
+    "ADDRESS2": ""
   }
